+++ src/utils/Dpi.h
@@ -4,10 +4,6 @@
 int DpiGetForHwnd(HWND);
 int DpiGetForMonitorOfHwnd(HWND hwnd);
 int DpiGet(HWND);
-void DbgLogDpi(const char* hypothesisId, const char* location, const char* message, HWND hwnd, int frameDpi,
-               int explicitDpi, int fontSize, int earlyReturn);
-void DbgLogFontMetrics(const char* hypothesisId, const char* location, const char* message, HWND hwnd);
-void DbgLogFontHandle(const char* hypothesisId, const char* location, const char* message, HFONT font, int wndDpi);
 int DpiScale(HWND, int);
 void DpiScale(HWND, int&, int&);